requirements.txt
Pillow
PyYAML
gdown
requests
tqdm